- more warnings

+++ copasi/MIRIAMUI/CQCreatorDM.cpp
@@ -1,4 +1,4 @@
-// Copyright (C) 2019 - 2021 by Pedro Mendes, Rector and Visitors of the
+// Copyright (C) 2019 - 2025 by Pedro Mendes, Rector and Visitors of the
 // University of Virginia, University of Heidelberg, and University
 // of Connecticut School of Medicine.
 // All rights reserved.
@@ -28,6 +28,8 @@
 #include "copasi/CopasiDataModel/CDataModel.h"
 #include "copasi/MIRIAM/CModelMIRIAMInfo.h"
 
+#include <cstddef>
+
 CQCreatorDM::CQCreatorDM(QObject *parent)
   : CQBaseDataModel(parent, NULL)
   , mpMIRIAMInfo(NULL)
@@ -51,7 +53,7 @@ size_t  CQCreatorDM::size() const
 
 int CQCreatorDM::rowCount(const QModelIndex& C_UNUSED(parent)) const
 {
-  return mFetched + 1;
+  return (int)mFetched + 1;
 }
 
 int CQCreatorDM::columnCount(const QModelIndex&) const
@@ -214,7 +216,7 @@ bool CQCreatorDM::removeRows(int position, int rows, const QModelIndex & parent)
   if (rows <= 0)
     return true;
 
-  beginRemoveRows(parent, position, std::min< int >(mFetched, position + rows) - 1);
+  beginRemoveRows(parent, position, std::min< int >((int)mFetched, position + rows) - 1);
 
   std::vector< const CCreator * > ToBeDeleted;
   ToBeDeleted.resize(rows);
